Fix grouped FP8 vector conversion deduction
Match the grouped FP8 conversion helper's element-count template parameter to Vec's uint32_t parameter so rowwise, columnwise, and activation instantiations can build.

Orchestra-Work-Order: wo_aea2e337b06582111bba66a6d6158a9e
Orchestra-Task: task_30c4b6ddb896e5ea3ca5b54731d2c819
Orchestra-Run: run_e95cdbb445943304622b95736f0eca49

+++ transformer_engine/common/cast/fp8/group_quantize_fp8.cuh
@@ -108,7 +108,7 @@ __device__ __forceinline__ void fast_scaled_fp8_cvt_4(const IType *input, OType 
 }
 
 template <bool IS_ACT, typename ParamOP, float (*OP)(float, const ParamOP &), typename IType,
-          typename OType, size_t NUM_ELTS>
+          typename OType, uint32_t NUM_ELTS>
 __device__ __forceinline__ void scaled_fp8_cvt_vec(const Vec<IType, NUM_ELTS> &input,
                                                    Vec<OType, NUM_ELTS> &output,
                                                    const size_t valid_cols, const float scale) {
